move error-boundary component to common folder
