+++ bigbluebutton-html5/imports/ui/components/common/error-boundary/component.jsx
