+++ bigbluebutton-html5/imports/ui/components/presentation-pod/container.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { withTracker } from 'meteor/react-meteor-data';
-import ErrorBoundary from '/imports/ui/components/error-boundary/component';
+import ErrorBoundary from '/imports/ui/components/common/error-boundary/component';
 import FallbackView from '/imports/ui/components/fallback-errors/fallback-view/component';
 import PresentationPodService from './service';
 import PresentationPods from './component';
