+++ bigbluebutton-html5/imports/ui/components/sidebar-content/component.jsx
@@ -9,7 +9,7 @@ import CaptionsContainer from '/imports/ui/components/captions/container';
 import BreakoutRoomContainer from '/imports/ui/components/breakout-room/container';
 import WaitingUsersPanel from '/imports/ui/components/waiting-users/container';
 import Styled from './styles';
-import ErrorBoundary from '/imports/ui/components/error-boundary/component';
+import ErrorBoundary from '/imports/ui/components/common/error-boundary/component';
 import FallbackView from '/imports/ui/components/fallback-errors/fallback-view/component';
 
 const propTypes = {
